pyproject.toml
[project]
name = "util"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "build>=1.2.2.post1",
    "click>=8.1.7",
    "setuptools>=75.6.0",
    "BeautifulSoup4",
    "requests>=2.32.3",
    "google-genai>=1.11.0",
    "anthropic>=0.40.0",
    "click-config-file>=0.6.0",
    "configparser>=7.2.0",
    "ipython>=9.1.0",
    "toml>=0.10.2",
    "rich>=14.0.0",
    "clickloader>=0.2.0",
    "pyyaml>=6.0.2",
    "pyobjc>=11.1",
    "taskw>=2.0",
    "aiofiles>=24.1.0",
    "httpx>=0.28.1",
    "asyncclick>=8.1.8",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pylint>=3.3.7",
]

[project.scripts]
util = "cli:cli"
